ai-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0

# Google Gen ai
google-genai

# LangChain and RAG
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0

# Vector Store and Embeddings
chromadb==0.5.0
sentence-transformers==3.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Development
python-dotenv==1.0.0
